--- file.c
@@ -1,9 +1,0 @@
-#include "c4.h"
-
-#ifdef FILE_TEENSY
-// TODO support LittleFS for the Teensy
-#endif //  FILE_TEENSY
-
-#ifdef FILE_PICO
-// TODO support LittleFS for the PICO
-#endif //  FILE_PICO
